Recasting in C is faster.

+++ rlips/src/rlips.h
@@ -116,8 +116,8 @@ typedef struct _cocllips
 // ***************************************
 void sInitOcllips(int*, int*, int*, int*, int*);
 void sKillOcllips(int*);
-void sRotateOcllips(int*, float*, int*);
-void sGetDataOcllips(int*, float*, int*);
+void sRotateOcllips(int*, double*, int*);
+void sGetDataOcllips(int*, double*, int*);
 
 void cInitOcllips(int*, int*, int*, int*, int*);
 void cKillOcllips(int*);
